Update CodeMirror CSS and Javascript files to version 3.15, under MIT-permissive license.
These files are exactly as they appear the upstream release 3.15 of
Codemirror, which was released under an MIT-permissive license.  To extract
these files, I did the following:

I downloaded the following file:
  http://codemirror.net/codemirror-3.15.zip
with sha256sum of:
$ sha256sum codemirror-3.15.zip
8cf3a512899852fd4e3833423ea98d34918cbf7ee0e4e0b13f8b5e7b083f21b9  codemirror-3.15.zip

And extracted from it the Javascript and CSS files herein committed, which
are licensed under the MIT-permissive license, placing them into their
locations in: rhodecode/public/{css,js}/

Using the procedure above, the only difference found between these files in
RhodeCode 2.2.5 release and herein were a few comments and whitespace.

Note that the file .../public/js/mode/meta_ext.js does *not* appear to be
part of CodeMirror and therefore is not included in this commit.