pull requests: allow closing of PRs in all states - undo #677
There can be reasons to close a PR without approving or rejecting it -
especially when approving or rejecting a PR overwrites the "review status" of
all changesets in it.

If some code for some reason is interested in PRs that has been closed as
approved or rejected then that is what they should look for ... even though
that is very hard with the current data model.