git: include an LF at the end of the service advertisement (Fixes #230)
This fixes hg-git/Dulwich and possibly other conservative Git clients,
which do not ignore the absence of the LF.

The original comment was a guess based on reverse engineering the protocol
not specified in the documentation yet at that moment. Now that the
documentation exists and states this explicitly, just do as it says.