tests: fix assertion rewriting in some tests with pytest-3.0.0+
Since pytest 3.0.0, asserts present in modules that are not directly seen by
pytest as 'test modules', are no longer rewritten to have improved
reporting, unless they are explicitly marked as up-for-rewriting.
Rationale from pytest upstream:

    However since we do not want to test different bytecode then you will
    run in production this hook only re-writes test modules themselves as
    well as any modules which are part of plugins. Any other imported module
    will not be re-written and normal assertion behaviour will happen.

This is e.g. the case for asserts in files like
kallithea/tests/api/api_base.py and kallithea/tests/models/common.py.

See
http://doc.pytest.org/en/latest/changelog.html#id13
http://doc.pytest.org/en/latest/writing_plugins.html#assertion-rewriting

This commit registers all modules below kallithea.tests for assert
rewriting, but note that asserts in kallithea/tests/__init__.py itself are
not rewritten as kallithea.tests is already imported when the register
statement is encountered.
Moving the register statement to kallithea/__init__.py would fix that, but
even then asserts in kallithea/tests/__init__.py seem not to be rewritten
(let alone the issue that we do not want a pytest dependency enforced in
kallithea/__init__.py which is also used in production).
Moving the code from kallithea/tests/__init__.py to a proper module, as
suggested by Mads Kiilerich, solves that problem.