style: put all datatable widgets on one line above the table
This makes tables look less messy - especially on the repogroups page where
there are 2 datatables.

We put all the info above the template so it is visible and stays in the same
place, no matter how many entries are shown or if filtering change the number.

TODO: We should probably get a wrapper around datatable instead of repeating
this so many times ...