config: move bulk of load_environment to new file app_cfg.py
In preparation of migrating to TurboGears2, where almost all setup code is
in one file app_cfg.py, move the existing Pylons-based configuration to the
same location. This will make it more clear which changes are really done
for TurboGears2.

The current move does not make functional changes, the code is moved
unmodified. Only real change is removal of some unused import statements.