i18n: remove old English translation file
Under TurboGears2, automatic user language detection based on browser settings
is enabled by default. If a user has 'English' as preferred language in the
browser, then the old English .mo file with a few dummy entries would be read
and used first, before trying the language configured in the Kallithea .ini
file (e.g. French). This renders a mix of languages in the UI ... which is
unexpected.

The hiding of configured translation is fixed by removing the old english
translationfile. It was preserved via commit 7d6c4bd58abd when removing all
other .mo files. The next release will require major config changes anyway, and
we don't try to be backwards compatible.