changelog: avoid cutting off commit messages
Split commit message into 2 <div> elements: one with just the first line, and
one with the full message. This makes it easier to handle the toggling of
showing the whole commit message. This way, we no longer have to mess with the
line height, and we avoid problems with cutting off parts of the text below the
text baseline.

Also stop cutting off the first line if it is too long for the table column.

It adds a bit of overhead to store the first line twice, but having them
separate make things simpler and is worth it.

Patch modified by Mads Kiilerich.