files: use same parent/child navigation as on changeset page
There are different reasons to do this. First, be consistent.
Second, steping through the revisions by revision number is problematic.
It means jumping between branches (I think even with the 'Follow current branch'
checkbox this is not obvious to all the users).
Also if we'd continue to use it, it'd require some special handling of hidden
revisions.