templates: disable paging and page size controls in DataTables
Since all data is sent to the client side, it can just as well be rendered.
PageDown or scrolling is more convenient for paging than paging buttons, so
short of "infinite scrolling", showing all entries is an acceptable solution
... especially when it very rarely is hundres of entries.

There could perhaps be further changes to how DataTables is shown - this is one
step.