less: simplify positioning of the +/- symbols on diff lines
As before, we place the +/- symbols in the pre:before so they default to the
same font and styling as the line content.

Instead of giving the pre a wide left margin and put the :before out there
using negative positing and visible overflow, just give the :before a fixed
width on all lines.

This change has been modified by Mads Kiilerich.