less: change markup and styling for the logo shown in the top left corner
Based on work by Dominik Ruf, but in this changeset, the goal is just to
improve markup and styling while keeping exactly the same look and feel. In
that process, it also fixes some browser specific issues that gave odd
spacing/positioning.

All styling (especially the logo image URL) is moved to CSS/LESS, using
background-image styling instead of an img tag.

The logo image is made a :before on the site branding text and will usually
have to align with this text, so we use inline-block to align to the baseline.

We set the right amount of negative margin-bottom so the image can go below the
baseline.

We use a "big enough" negative margin-top to allow the image to bleed outside
the line height and into the 12px top padding.

We assume the background image doesn't need cropping or panning and do thus not
set background-size, background-position, or background-repeat.