templates: mark site as IE10+ compatible to bypass Compatibility mode
Internet Explorer 11 (and probably other versions) seems to ship with the
setting "Display intranet sites in Compatibility View" enabled by default.
This compatibility view means that modern elements of CSS and JavaScript are
not (well) support and makes Kallithea look very bad and become unusable.

This is a problem when hosting Kallithea in a corporate environment where
its URL is detected as 'the intranet'.

Solve the issue by explicitly marking Kallithea to be compatible with IE10,
so that Internet Explorer 10 and above will show Kallithea correctly in
Intranet environments.