tests: api: cover original behavior for 'pull'
Up to commit 85d812ab4c64, the api 'pull' method can either pull from the
repository's specified remote (clone_uri) or, if the repository is a fork,
from the fork origin.

The existing test case only covered pulling from a remote, and pulling from
a repo (non-fork) that does not have a remote specified.

Update the test method naming to clarify this, and add a test to pull from a
fork.