issues: make issue_prefix optional again
Commit 39a59e6915bb398b42c3c2a63c48a950e9d63b55 (helpers: refactor and
optimize urlify_issues) made issue_prefix mandatory, while previously it
could be empty. An empty issue_prefix is useful when the entire issue
pattern needs to be used in the created link.

For example, consider a pattern 'PR123' that needs to be translated into:
http://example.com/pullrequests/PR123.

This could be configured with:
    issue_pat = (PR\d+)
    issue_server_link = http://example.com/pullrequests/{id}
    issue_prefix =

We still refuse the issue pattern when issue_prefix is not present at all.