issues: add more tests and restructure
Add two test cases using the default issue pattern from the ini file
template prior to commit 9cef5615da7b, and some more cases, some of which
verify the behavior regarding test cases with mandatory/optional leading
whitespace.

Increase readability by grouping tests with the same issue pattern, server
and prefix, and place the urlified result on a separate line from the input,
wrapped in case there are multiple issues in the input.

For the result, """ string delimiters are used instead of ' as used in the
input, to visually distinguish them better.