less: fix changelog commit message flowing into the labels that are floating to the right of it
For long commit messages, two texts were shown on top of each other and none of
them could be read.

To fix that, give the floating labels a solid white background.

To avoid a hard cut-off of the commit message, use a white box-shadow to the
left of the labels to give a smooth transition.

To avoid the commit message flowing into the td padding to the right of the
labels, make log-container hide overflow.