make-release: import version and copyright updates from default branch (dba4e770d4b6)
make-release has been improved on the default branch in commit dba4e770d4b6,
and that version of the script also almost works on stable.

There are some parts we don't have in stable branch: dev_requirements.txt,
run-all-cleanup, and a working dist file content check:

The MANIFEST.in file on the stable branch is not complete.
Most of the missing entries could simply be added
(.travis.yml, requirements.txt, tox.ini, and scripts/*) but there is one
problematic entry: kallithea/i18n/en/LC_MESSAGES/kallithea.mo.
In the check command, all .mo entries are filtered out, causing a remaining
diff for the English .mo.

As that file has been removed on the default branch already, and the missing
entries are not important for the release anyway, simply drop the check
for now.