setup-db: perform an initial repository scan as stated by the docs (Issue #302)
The documentation, as well as the prompt text inside setup-db itself, states
that the repository root location will be scanned automatically for existing
repositories.
However, this is not actually the case. Only exception is when the
'initial_repo_scan' is set to True in the ini file; in that case the scan is
done on each start of Kallithea.

Add the required repo scan logic at the end of setup-db, after setting up
the application completely (the database has only just been set up).

The app setup code from BasePasterCommand is duplicated - this command doesn't
fit in and need to run both before and after database creation.

The scan call from kallithea/config/app_cfg.py is also duplicated - that will
be removed next.