style: use monospace on all multi-line form inputs
Multi-line form input (textarea) is used for PR comments, PR descriptions,
repository descriptions and 'HTML/CSS/Javascript customization block'.
In the first three cases, the rendered result is displayed in monospace, so
the corresponding textarea should be as well.
As the expected content of the HTML/CSS/Javascript block is code, it makes
perfect sense to use monospace there too.