front-end: Use select2 from node_modules and stop bundling it
select2 3.5.4 was added in 304e83e9bcde ... but the latest npm release in the 3
series is 3.5.1, so we use that one instead. We should probably upgrade to the
4 series.

The select2 images were not in the location the generated css pointed - now we
copy them from node_modules to the right location, next to the generated css.

Note: this will drop 190cb30841de "branches: fix performance of branch
selectors with many branches - only show the first 200 results" ... but it
should no longer be relevant now when we use server side filtering.

15e507047bae introduced select2-bootstrap.css - it is not clear what version
was used, but we use the latest 1.4.6 which also is very old.