files: use the web browsers built-in js history instead of native.history.js
The history API is available in all web browsers we support.

window.history.pushState is called to register a state that we can go
back/forward to. (But contrary to native.history.js, it doesn't do any
immediate processing of the state and doesn't actually navigate to it.)

When navigation occurs, we get the popstate event and invoke load_state to
actually load the state.