tests: fix assert rewriting in non-test modules like api_base.py
pytest rewrites assert statements in tests so it can print details about the
values involved when the assert fails.

Since pytest 3.0.0, this was no longer the case for files/modules that are
not discovered as test modules, i.e. starting with 'test_'. Examples are:
    api/api_base.py
    models/common.py
    base.py
    fixture.py

Commit 790aeeddcab598d2aacbbecf86830b608ca8b32b attempted to fix that
problem, but seems incorrect. A trailing dot should not be there.