cli: fill in git_hook_interpreter at 'config-create' time to really fix potentially invalid interpreter in git hooks (Issue #333)
When generating a configuration file using 'kallithea-cli config-create', it is
probably using the right Python interpreter, so fill in the current Python
executable as 'git_hook_interpreter' in the generated file.

There should thus rarely be any need to configure this manually, and issue #333
will *really* be fixed.

As this causes an absolute path to be encoded inside the ini file, moving
the virtualenv will require updating this path.

For development.ini we do not want to hardcode any path and are happy to leave
it using the old heuristics at runtime.

Suggested-by: Mads Kiilerich <mads@kiilerich.com>