setup: bump some setup.py dependency versions
Upgrade notes for these libraries have not been investigated thoroughly, but
testing seems to show that it works. We are also early in the development
phase, so big problems will be caught by general testing before going wide.

Note: TurboGears2 is not upgraded to 2.4 yet. That upgrade would require us to
first move away from using the Pylons compatibility layer.