i18n: disable 'no-wrap' on extract_messages to align wrap settings with weblate
In Weblate, the default wrap settings are used and this is not configurable.
This means that .po files will have wrapping at 76 characters.

On the other hand, the 'extract_messages' method in Kallithea was configured
to not wrap at all. When regenerating .po files based on a new .pot file,
there could thus be wrapping changes, back and forth.

Avoid this by removing the 'no-wrap' setting and falling back to (hopefully)
the same default as Weblate.

Instead of:

  msgid "very long line"

it will now emit it more readable:

  msgid ""
  "very long "
  "line"