auth: make crowd logging simpler
There is no point in creating dicts and then logging them as json. Also, json
can't handle py3 bytes and it would fail on py3. (ext_json could perhaps handle
bytes, but it seems better to keep it simple and explicit.)

If the default repr isn't good enough, it would be better to use pprint. But
repr is good enough.