ui: show toggleable "Follow" status in repo groups' repo list
It makes sense to show Follow status next to repo names in the repo list, and
it is a meaningful and efficient bulk operation to toggle Follow status there.

Clicking on the (Un)Follow 'heart' will toggle the caller's follow status for
that repo.

The repo model already has layering violations - expand on them to compute the
follow status of the current user.

(Changeset was cherry picked and modified by Mads Kiilerich.)