scripts/i18n: introduce new i18n maintenance script
The translation files in the Kallithea repository contained references to
the location(s) of each string in the repository. This is useful to
translators, but is not needed for all other users.

The big problem with that information is that it changes very commonly as a
result of normal development in Kallithea, causing a lot of unimportant
delta in the Kallithea repository, thus causing unnecessary repository
growth.

A script 'i18n' is added to help maintain the i18n files. Functionality will be
added later.