auth: for default permissions, use existing explicit query result values instead of following dot references in ORM result objects
There has been reports of spurious crashes on resolving references like
.repository from Permissions:

  File ".../kallithea/lib/auth.py", line 678, in __wrapper
    if self.check_permissions(user):
  File ".../kallithea/lib/auth.py", line 718, in check_permissions
    return user.has_repository_permission_level(repo_name, self.required_perm)
  File ".../kallithea/lib/auth.py", line 450, in has_repository_permission_level
    actual_perm = self.permissions['repositories'].get(repo_name)
  File ".../kallithea/lib/vcs/utils/lazy.py", line 41, in __get__
    value = self._func(obj)
  File ".../kallithea/lib/auth.py", line 442, in permissions
    return self.__get_perms(user=self, cache=False)
  File ".../kallithea/lib/auth.py", line 498, in __get_perms
    return compute(user_id, user_is_admin)
  File ".../kallithea/lib/auth.py", line 190, in _cached_perms_data
    r_k = perm.UserRepoToPerm.repository.repo_name
  File ".../sqlalchemy/orm/attributes.py", line 285, in __get__
    return self.impl.get(instance_state(instance), dict_)
  File ".../sqlalchemy/orm/attributes.py", line 721, in get
    value = self.callable_(state, passive)
  File ".../sqlalchemy/orm/strategies.py", line 710, in _load_for_state
    % (orm_util.state_str(state), self.key)

  sqlalchemy.orm.exc.DetachedInstanceError: Parent instance <UserRepoToPerm at ...> is not bound to a Session; lazy load operation of attribute 'repository' cannot proceed (Background on this error at: http://sqlalche.me/e/bhk3)

Permissions are cached between requests: SA result records are stored in in
beaker.cache.sql_cache_short and resued in following requests after the initial
session as been removed. References in Permission objects would usually give
lazy lookup ... but not outside the original session, where we would get an
error like this.

Permissions are indeed implemented/used incorrectly. That might explain a part
of the problem. Even if not fully explaining or fixing this problem, it is
still worth fixing:

Permissions are fetched from the database using Session().query with multiple
class/table names (joined together in way that happens to match the references
specified in the table definitions) - including Repository. The results are
thus "structs" with selected objects. If repositories always were retrieved
using this selected repository, everything would be fine. In some places, this
was what we did.

But in some places, the code happened to do what was more intuitive: just use
.repository and rely on "lazy" resolving. SA was not aware that this one
already was present in the result struct, and would try to fetch it again. Best
case, that could be inefficient. Worst case, it would fail as we see here.

Fix this by only querying from one table but use the "joinedload" option to
also fetch other referenced tables in the same select. (This might
inefficiently return the main record multiple times ... but that was already
the case with the previous approach.)

This change is thus doing multiple things with circular dependencies that can't
be split up in minor parts without taking detours:

The existing repository join like:
        .join((Repository, UserGroupRepoToPerm.repository_id == Repository.repo_id))
is thus replaced by:
        .options(joinedload(UserGroupRepoToPerm.repository))

Since we only are doing Session.query() on one table, the results will be of
that type instead of "structs" with multiple objects. If only querying for
UserRepoToPerm this means:
  - perm.UserRepoToPerm.repository becomes perm.repository
  - perm.Permission.permission_name looked at the explicitly queried Permission
    in the result struct - instead it should look in the the dereferenced
    repository as perm.permission.permission_name