diff: allow commenting on meta lines (Issue #329)
There is no good reason it only should be possible to comment on content lines.
Other lines might not have an obvious locator, but we can live with that as
long as each comment only apply in one place.

With this, we actually want the comment bubble on all lines with bubble markup,
so we can loosen the css selector.