vcs: fix cloning remote repository with HTTP authentication (Issue #379)
Using a remote clone URI of
   http://user:pass@host/...
triggered an exception:

    ...
    E     File ".../kallithea/lib/utils.py", line 256, in is_valid_repo_uri
    E       GitRepository._check_url(url)
    E     File ".../kallithea/lib/vcs/backends/git/repository.py", line 183, in _check_url
    E       passmgr.add_password(*authinfo)
    E     File "/usr/lib/python3.7/urllib/request.py", line 848, in add_password
    E       self.reduce_uri(u, default_port) for u in uri)
    E     File "/usr/lib/python3.7/urllib/request.py", line 848, in <genexpr>
    E       self.reduce_uri(u, default_port) for u in uri)
    E     File "/usr/lib/python3.7/urllib/request.py", line 875, in reduce_uri
    E       host, port = splitport(authority)
    E     File "/usr/lib/python3.7/urllib/parse.py", line 1022, in splitport
    E       match = _portprog.fullmatch(host)
    E   TypeError: cannot use a string pattern on a bytes-like object


The authinfo tuple is obtained via mercurial.util.url, which unfortunately
returns a tuple of bytes whereas urllib expects strings.
It seems that mercurial internally has some more hacking around urllib as
urllibcompat.py, which we don't use.

Therefore, transform the bytes into strings before passing authinfo to
urllib. As the realm can be None, we need to check it specifically otherwise
safe_str would return a string 'None'.

A basic test that catches the mentioned problem is added, even though it
does not actually test that cloning with auth info will actually work (it
only tests that it fails cleanly if the URI is not reachable).

Additionally, one use of 'test_uri' in hg/repository.py still needed to be
transformed from bytes to string. For git this was already ok.