model: handle redundant reviewers in add_reviewers
Don't attempt to add reviewers that are already a reviewer for the specified
PR (redundant reviewers).

Return the list of added and redundant reviewers, for the controller to
handle.

Under normal circumstances, the pullrequest controller already processes the
list of reviewers and only calls add_reviewers for new reviewers. But, there
could be ways were this checking fails, for example due to a race condition
between two simultaneous requests for the same pullrequest, or due to a bug
in the web server framework that causes the same request to be handled
again.