lib/diffs: make sure that trailing tabs are indicated
Between the initial submission and final version of commit f79c40759d6f,
changes were made that turn out to be incorrect. The changes assume that the
later match on trailing tabs will 'win' from the plain 'tab' match. However,
Python 're' documentation says:

    As the target string is scanned, REs separated by '|' are tried from
    left to right. When one pattern completely matches, that branch is
    accepted. This means that once A matches, B will not be tested further,
    even if it would produce a longer overall match. In other words, the '|'
    operator is never greedy.
    https://docs.python.org/3.8/library/re.html

As a result, a trailing tab is seen as a plain tab and not highlighted in a
special way.

Unify the tab handling to make it unambiguous how they should be parsed.

The change diff mainly shows re group numbers shifting.