usergroup: fix adding permission
Clicking "Add new" on /_admin/user_groups/1/edit/perms would flicker the input
fields, and then show "User group permissions updated".

That turns out to be because HTML button tags default to 'submit' inside forms,
and in this place (but not for user or repo group permissions) we didn't
specify the button type as 'button'.

There might be other similar problems.