setup: bump max versions of pip package dependencies
It is a bit risky to update versions on stable, but it is also risky to stay on
old versions when there are better(?) options.

There are various reasons these max versions have been chosen. I have not
analyzed problems seen when using higher versions, but the versions specified
here seems like a safe progression of the broad baseline.