setup: bump maximum version of Mercurial to 7.0.x
The freeze for the Mercurial 7.0 release was done on February 28th. The release
of 7.0.0 was scheduled for March 1st, but was delayed. It is likely that the
release will happen soon without changes that would break Kallithea.

The tests pass and some quick manual testing of the web UI found no issues.